move fixRtl to Rtlutils
